requirements.txt
xsix==1.15.0
tensorflow==2.6.0
streamlit==1.0.0
numpy==1.19.5
opencv-python-headless
Pillow==8.3.2
